% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{available_ranks}
\alias{available_ranks}
\title{Get a Table of Available Ranks}
\usage{
available_ranks()
}
\value{
a tibble with columns \code{level} and \code{de} (ranks in German).
}
\description{
Get the available ranks as a data frame. The ranks are returned in the
correct hierarchical order from highest to lowest. Other ranks may be used,
but only those listed here will be represented by coloured nodes in the
visualisation.
}
