% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{count_ranks}
\alias{count_ranks}
\title{Count the Number of Taxa per Rank}
\usage{
count_ranks(graph, subgraph = NULL, by_rank = NULL)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object}

\item{subgraph}{the name of a taxon which is to be used as the root of a
subgraph, to which counting is restricted.}

\item{by_rank}{the name of a rank. Ranks will be counted for each of the
taxa with this rank.}
}
\value{
a tibble
}
\description{
Count the number of taxa per rank in a \code{taxonomy_graph} and return
the result as a tibble.
}
\details{
If \code{by_rank} is not used, the number of taxa is counted in the complete
graph or in the subgraph specified by \code{subgraph}.

\code{by_rank} can be used to count ranks grouped by the taxa of a given rank.
In this case, each row will correspond to a taxon of the requested rank
and each column will correspond to a rank.
}
\examples{
file <- get_example_taxonomy_file()
taxonomy <- read_taxonomy(file)

# count the complete graph
count_ranks(taxonomy)

# count only the cats
count_ranks(taxonomy, subgraph = "Katzen")

# count caniformia by family
count_ranks(taxonomy, "Hundeartige", by_rank = "Familie")

}
