% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{enrich_taxonomy_with_images}
\alias{enrich_taxonomy_with_images}
\title{Enrich a Taxonomy File with URLs to Wikipedia-Images}
\usage{
enrich_taxonomy_with_images(
  file,
  delim = ",",
  contact = NULL,
  retry = FALSE,
  progress = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{file}{path to the csv file}

\item{delim}{the delimiter used in the file}

\item{contact}{contact data to pass on in the user agent of the API call to
Wikipedia. Starting in 2026, there is a strict rate limit on API calls per
minute for calls without contact data. There are two possible ways to pass
contact data:
\itemize{
\item a string containing an email address.
\item a list that specifies an existing Wikipedia account. You need to specify two
named elements: "lang" giving the language code of the Wikipedia where the
account is registered and "user" giving the user name.
Example: \code{list(lang = "de", user = "James Bond")}.
}

See the \href{https://foundation.wikimedia.org/wiki/Policy:Wikimedia_Foundation_User-Agent_Policy}{User Agent Policy}
for more information.}

\item{retry}{Whether to retry images that have not been found previously.}

\item{progress}{Whether to show a progress bar.}

\item{quiet}{Should all output be suppressed?}
}
\value{
a \code{taxonomy_graph} with added image URLs. The file given by \code{file} is
overwritten as a side effect.
}
\description{
Add URLs to Images from Wikipedia to a file with a taxonomic hierarchy.
By default, existing URLs are kept and only missing URLs are filled in.
}
