% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_taxon}
\alias{find_taxon}
\title{Find Taxa By Pattern}
\usage{
find_taxon(graph, pattern, target = c("all", "name", "scientific"))
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object, typically created with
\code{\link[=read_taxonomy]{read_taxonomy()}}.}

\item{pattern}{a regex pattern, matching is case insensitive.}

\item{target}{character giving the column to search in
("name" or "scientific"). The default is "all" which searches in all columns.}
}
\value{
a character vector giving the taxon names that match the pattern
}
\description{
Find taxa by pattern and return their common names. The output can be passed
to the arguments \code{show}, \code{full_expand} or \code{focus} in \code{\link[=plot_taxonomy]{plot_taxonomy()}}.
}
\examples{
file <- get_example_taxonomy_file()
taxonomy <- read_taxonomy(file)
find_taxon(taxonomy, "katze")
find_taxon(taxonomy, "felis")

}
