% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_example_taxonomy_file}
\alias{get_example_taxonomy_file}
\title{Get Example Taxonomy File}
\usage{
get_example_taxonomy_file()
}
\description{
Return the path to an example file containing a partial taxonomic hierarchy
of the order Carnivora.
It can be read with \code{\link[=read_taxonomy]{read_taxonomy()}} and plotted with \code{\link[=plot_taxonomy]{plot_taxonomy()}}.
}
