% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{get_parent_taxon}
\alias{get_parent_taxon}
\title{Get the Parent Taxon of a Taxon}
\usage{
get_parent_taxon(graph, taxon)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object, typically created with
\code{\link[=read_taxonomy]{read_taxonomy()}}.}

\item{taxon}{character of length 1 giving the exact common or scientific name
of a taxon.}
}
\value{
a character vector of length 1 giving the label of the parent taxon of
\code{taxon}, except when \code{taxon} is the root, in which case an empty character
vector is returned.
}
\description{
Get the Parent Taxon of a Taxon
}
