% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{get_subgraph}
\alias{get_subgraph}
\title{Extract a Subgraph from a Taxonomy Graph}
\usage{
get_subgraph(graph, taxon)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object}

\item{taxon}{character giving a single taxon. The subgraph will contain
this taxon as the root and all the taxa that are in the tree below it.}
}
\value{
a \code{taxonomy_graph} object
}
\description{
Extract a subgraph from a \code{taxonomy_graph} below a given taxon.
}
\examples{
file <- get_example_taxonomy_file()
taxonomy <- read_taxonomy(file)
cats <- get_subgraph(taxonomy, "Katzen")

}
