% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{get_taxon_names}
\alias{get_taxon_names}
\title{Get Taxon Labels by Common or Scientific Name}
\usage{
get_taxon_names(graph, char)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object, typically created with
\code{\link[=read_taxonomy]{read_taxonomy()}}.}

\item{char}{a character vector of common and/or scientific names to be
converted to labels. Matching is case-sensitive.}
}
\value{
a character vector of the same length as \code{char} containing taxon labels.
}
\description{
The Taxa in the \code{taxonomy_graph} are labelled by their common name.
This function returns the label for a vector of common and/or scientific
names. It is much more efficient than \code{\link[=find_taxon]{find_taxon()}} because it relies
on a matching table stored in the \code{taxonomy_graph}.
}
\seealso{
find_taxon
}
