% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{get_tree_depth}
\alias{get_tree_depth}
\title{Get Depth of a Taxonomy Tree Graph}
\usage{
get_tree_depth(graph)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object}
}
\value{
integer giving the depth of the graph
}
\description{
Get the depth of a taxonomy tree graph, i.e., the number of nodes from the
root to the most distant leaf.
}
\examples{
file <- get_example_taxonomy_file()
taxonomy <- read_taxonomy(file)
get_tree_depth(taxonomy)

}
