% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_wikipedia_image_urls}
\alias{get_wikipedia_image_urls}
\title{Download URL for Wikipedia Images}
\usage{
get_wikipedia_image_urls(
  taxa,
  contact = NULL,
  size = 120,
  lang = "de",
  progress = TRUE
)
}
\arguments{
\item{taxa}{character with the taxa for which an image URL should be
obtained.}

\item{contact}{contact data to pass on in the user agent of the API call to
Wikipedia. Starting in 2026, there is a strict rate limit on API calls per
minute for calls without contact data. There are two possible ways to pass
contact data:
\itemize{
\item a string containing an email address.
\item a list that specifies an existing Wikipedia account. You need to specify two
named elements: "lang" giving the language code of the Wikipedia where the
account is registered and "user" giving the user name.
Example: \code{list(lang = "de", user = "James Bond")}.
}

See the \href{https://foundation.wikimedia.org/wiki/Policy:Wikimedia_Foundation_User-Agent_Policy}{User Agent Policy}
for more information.}

\item{size}{integer giving the width of the image in pixel.}

\item{lang}{language of the Wikipedia page to use. Available languages
include "de" (German, the default), "en" (English), "fr" (French), and
"es" (Spanish). See
\href{https://meta.wikimedia.org/wiki/List_of_Wikipedias#All_Wikipedias_ordered_by_number_of_articles}{List of Wikipedias}
for a full list of available languages.}

\item{progress}{Whether to show a progress bar.}
}
\value{
a character vector with URLs. For taxa where no thumbnail was found, the
function returns \code{NA}.
}
\description{
Wikipedia pages usually have a thumbnail associated with them. This function
uses an API offered by Wikipedia to extract the URL of that image.
}
