% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_taxonomy}
\alias{plot_taxonomy}
\title{Plot a Taxonomic Hierarchy}
\usage{
plot_taxonomy(
  graph,
  show = c(),
  expand_rank = c(),
  full_expand = c(),
  focus = c(),
  highlight = c(),
  show_images = FALSE,
  image_size = c("250", "60", "120", "330", "500"),
  link_length = 150,
  font_size = 12,
  highlight_missing_images = FALSE,
  local_country = NULL
)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object, typically created with
\code{\link[=read_taxonomy]{read_taxonomy()}}.}

\item{show}{character giving the common or scientific names of taxa that
should be visible.
The tree will be shown uncollapsed up to those taxa.}

\item{expand_rank}{character giving the names of ranks that should always
be expanded.}

\item{full_expand}{character giving the common or scientific names of
taxa that should be fully
expanded, i.e., all taxa below the given taxon should be visible. Note
that this does not expand the graph above the given taxa, such that the
expanded part may be invisible. Use \code{show} to expand the graph up to a
given taxon.}

\item{focus}{character giving one or several taxa to focus on. This means
that the tree up to this taxon will be expanded as well as the full
subtree below that taxon. It is equivalent to putting the same taxa in
\code{show} and \code{full_expand}. If \code{focus} is used, those other two arguments will
be ignored.}

\item{highlight}{character giving the common or scientific names of taxa
that should be
highlighted.}

\item{show_images}{should images be shown in tooltips. This requires that
image URLs are contained in the \code{taxonomy_graph}. URLs from Wikipedia can
be automatically added using \code{\link[=enrich_taxonomy_with_images]{enrich_taxonomy_with_images()}}.}

\item{image_size}{integer giving the width of images in pixels.}

\item{link_length}{length of the horizontal links that connect nodes
in pixels.}

\item{font_size}{font size of the labels in pixels.}

\item{highlight_missing_images}{logical, should nodes with missing
image link by highlighted?}

\item{local_country}{country code for the "local country". If the taxonomy
has taxa that are marked as local, the flag of this country will be
shown in the tooltip. If the columns "local" does not exist in the
data, this argument has no effect. If no country code is given, a
generic location marker will be used instead.}
}
\value{
a \code{collapsibleTree} htmlwidget with the interactive tree visualisation.
}
\description{
Create an interactive visualisation of a taxonomic hierarchy.
}
\details{
The function makes use of the html widget defined in the package
\link[collapsibleTree:collapsibleTree]{collapsibleTree}
but not of the function
\code{\link[collapsibleTree:collapsibleTreeNetwork]{collapsibleTreeNetwork()}}
from that package that can produce the same plot from the same input data.
\code{plot_taxonomy()} is less flexible than the function from
collapsibleTree but also much faster.
}
\examples{
file <- get_example_taxonomy_file()
taxonomy <- read_taxonomy(file)
plot_taxonomy(taxonomy)

# expand the full path to the grizzly bear
plot_taxonomy(taxonomy, show = "Grizzlybär")

# fully expand the cat family (Felidae)
# The cats are not visible, because the nodes above are not expanded.
plot_taxonomy(taxonomy, full_expand = "Katzen")

# expand the path up to the family of the cats and everything below
plot_taxonomy(taxonomy, focus = "Katzen")

# add images to tooltips
plot_taxonomy(taxonomy, focus = "Bären", show_images = TRUE)

# show and highlight black bear and kodiak bear
plot_taxonomy(taxonomy,
              show = c("Amerikanischer Schwarzbär", "Kodiakbär"),
              highlight = c("Amerikanischer Schwarzbär", "Kodiakbär"))

}
