% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_taxonomy}
\alias{run_taxonomy}
\title{Run simpleTaxonomy Shiny App}
\usage{
run_taxonomy(
  file = NULL,
  root = NULL,
  expand_ranks = c("Gattung", "Art", "Unterart"),
  image_size = c("250", "60", "120", "330", "500"),
  link_length = 200,
  local_country = NULL,
  launch_browser = NULL
)
}
\arguments{
\item{file}{Path to a file containing a \code{taxonomy_graph}. If omitted, the
data from \url{https://github.com/stibu81/taxonomyData} is used.}

\item{root}{character giving the common or scientific name of the taxon to
select as root. If omitted, the actual root of the taxonomy graph is used.
If the selected root does not exist in the graph, the app also falls
back on the actual root of the graph.}

\item{expand_ranks}{a list of ranks that should always be expanded. This
can be changed interactively in the app.}

\item{image_size}{numeric giving the default image width in pixels used in
the app. The user can change this interactively in the app.}

\item{link_length}{numeric giving the default length of links in pixels
used in the app. The user can change this interactively in the app.}

\item{local_country}{country code for the "local country". If the taxonomy
has taxa that are marked as local, the flag of this country will be
shown in the tooltip. If the columns "local" does not exist in the
data, this argument has no effect. If no country code is given, a
generic location marker will be used instead.}

\item{launch_browser}{logical, if \code{TRUE}, the application
is opened in the system's default browser, if \code{FALSE},
no browser is started. If the argument is omitted, the value
according to the option \code{shiny.launch.browser} is used,
which in RStudio opens the internal shiny viewer.}
}
\description{
Run the simple shiny app that is included with simpleTaxonomy. The app
provides several features to interact with a \code{taxonomy_graph}.
A \code{taxonomy_graph} can be read from a file or an URL. By default, the
app loads the data from \url{https://github.com/stibu81/taxonomyData}.
}
