% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_taxonomy}
\alias{save_taxonomy}
\title{Save Taxonomic Hierarchy to HTML-File}
\usage{
save_taxonomy(widget, file, background = "white", title = "simpleTaxonomy")
}
\arguments{
\item{widget}{a widget created with \code{\link[=plot_taxonomy]{plot_taxonomy()}}.}

\item{file}{the full path to the file}

\item{background}{character giving the html background colour of
the widget. Defaults to white.}

\item{title}{text to use as the title of the saved page}
}
\description{
Save a plot of a taxonomic hierarchy created with \code{\link[=plot_taxonomy]{plot_taxonomy()}} to
a self-contained HTML-file.
}
