% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{enrich_taxonomy_with_images}
\alias{enrich_taxonomy_with_images}
\title{Enrich a Taxonomy File with URLs to Wikipedia-Images}
\usage{
enrich_taxonomy_with_images(file, delim = ",")
}
\arguments{
\item{file}{path to the csv file}

\item{delim}{the delimiter used in the file}
}
\value{
a \code{taxonomy_graph} with added image URLs. The file given by \code{file} is
overwritten as a side effect.
}
\description{
Add URLs to Images from Wikipedia to a file with a taxonomic hierarchy.
By default, existing URLs are kept and only missing URLs are filled in.
}
