% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flag_icon}
\alias{flag_icon}
\title{Create an svg Flag Icon Tag}
\usage{
flag_icon(code)
}
\arguments{
\item{code}{character with the flag code, e.g., \code{"ch"} or \code{"de"}.
To find available flag codes, use \code{\link[=get_flag_info]{get_flag_info()}}.}
}
\description{
Create an html \verb{<img>} tag for a country flag svg file that is shipped with
the package. The flag icons are by \href{https://github.com/lipis}{Lipis} and were
downloaded from \url{https://github.com/lipis/flag-icons}.
}
