% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{get_flag_info}
\alias{get_flag_info}
\title{List Information About the Available Flags}
\usage{
get_flag_info(filter = NULL)
}
\arguments{
\item{filter}{a regex pattern that is used to search in the country names.
Search is case-insensitive.}
}
\value{
a tibble with character columns \code{name}, \code{continent}, \code{code}, and \code{capital}.
}
\description{
Return a table that for all available flags returns the country name,
the continent, the flag code and the capital. The flag code must be used
as input in functions like \code{\link[=flag_icon]{flag_icon()}} and \code{\link[=run_taxonomy]{run_taxonomy()}}.
}
