% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{get_root_node}
\alias{get_root_node}
\title{Get the Root Node of a Taxonomy Graph}
\usage{
get_root_node(graph)
}
\arguments{
\item{graph}{a \code{taxonomy_graph} object}
}
\value{
a \code{igraph.vs} object corresponding to the root node of the graph.
}
\description{
Get the Root Node of a Taxonomy Graph
}
