% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_wikipedia_image_urls}
\alias{get_wikipedia_image_urls}
\title{Download URL for Wikipedia Images}
\usage{
get_wikipedia_image_urls(taxa, size = 100, progress = TRUE)
}
\arguments{
\item{taxa}{character with the taxa for which an image URL should be
obtained.}

\item{size}{integer giving the width of the image in pixel.}

\item{progress}{Wheter to show a progress bar.}
}
\value{
a character vector with URLs. For taxa where no thumbnail was found, the
function returns \code{NA}.
}
\description{
Wikipedia pages usually have a thumbnail associated with them. This function
uses an API offered by Wikipedia to extract the URL of that image.
}
