% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_taxonomy}
\alias{plot_taxonomy}
\title{Plot a Taxonomic Hierarchy}
\usage{
plot_taxonomy(data, find = c(), font_size = 12)
}
\arguments{
\item{data}{tibble defining the taxonomic hierarchy, typically created with
\code{\link[=read_taxonomy]{read_taxonomy()}}.}

\item{find}{character giving the names of taxons that should be visible.
The tree will be shown uncollapsed up to those taxons.}

\item{font_size}{integer giving the font size of the labels in pixels.}
}
\value{
a htmlwidget with the interactive tree visualisation.
}
\description{
Create an interactive visualisation of a taxonomic hierarchy.
}
