% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_taxonomy}
\alias{plot_taxonomy}
\title{Plot a Taxonomic Hierarchy}
\usage{
plot_taxonomy(
  data,
  show = c(),
  expand_rank = c(),
  full_expand = c(),
  focus = c(),
  font_size = 12
)
}
\arguments{
\item{data}{tibble defining the taxonomic hierarchy, typically created with
\code{\link[=read_taxonomy]{read_taxonomy()}}.}

\item{show}{character giving the names of taxons that should be visible.
The tree will be shown uncollapsed up to those taxons.}

\item{expand_rank}{character giving the names of ranks that should always
be expanded.}

\item{full_expand}{character giving the names of taxons that should be fully
expanded, i.e., all taxons below the given taxon should be visible. Note
that this does not expand the graph above the given taxon, such that the
expanded part may be invisible. Use \code{show} to expand the graph up to a
given taxon.}

\item{focus}{character giving one or several taxons to focus on. This means
that the tree up to this taxon will be expanded as well as the full
subtree below that taxon. It is equivalent to putting the same taxons in
\code{show} and \code{full_expand}. If \code{focus} is used, those other two arguments will
be ignored.}

\item{font_size}{integer giving the font size of the labels in pixels.}
}
\value{
a htmlwidget with the interactive tree visualisation.
}
\description{
Create an interactive visualisation of a taxonomic hierarchy.
}
