% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTaxonomy-package.R
\docType{package}
\name{simpleTaxonomy-package}
\alias{simpleTaxonomy}
\alias{simpleTaxonomy-package}
\title{simpleTaxonomy: Create Interactive Visualisation of Taxonomic Hierarchy}
\description{
Create interactive visualisations of a taxonomy hierarchy that is specified in a csv file using a simple format. The visualisation uses the javascript library D3.js via the R package collapsibleTree.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stibu81.github.io/simpleTaxonomy/}
  \item Report bugs at \url{https://github.com/stibu81/simpleTaxonomy/issues}
}

}
\author{
\strong{Maintainer}: Stefan Lanz \email{slanz1137@gmail.com}

}
\keyword{internal}
